Unify sub-modules into output.py, as suggested for GHI #2
