--- piplicenses/output/json.py
@@ -1,84 +1,0 @@
-# pip-licenses-cli
-#
-# MIT License
-#
-# Copyright (c) 2018 raimon
-# Copyright (c) 2025 stefan6419846
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in
-# all copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
-from __future__ import annotations
-
-from typing import TYPE_CHECKING
-
-from prettytable import PrettyTable
-
-if TYPE_CHECKING:  # pragma: no cover
-    from prettytable import RowType
-
-
-class JsonPrettyTable(PrettyTable):
-    """PrettyTable-like class exporting to JSON"""
-
-    def format_row(self, row: RowType) -> dict[str, str | list[str]]:
-        resrow: dict[str, str | list[str]] = {}
-        for field, value in zip(self._field_names, row):
-            resrow[field] = value
-
-        return resrow
-
-    def get_string(self, **kwargs: str | list[str]) -> str:
-        # import included here in order to limit dependencies
-        # if not interested in JSON output,
-        # then the dependency is not required
-        import json
-
-        # Fetch formatting options (e.g., selected fields, row slice)
-        options = self._get_options(kwargs)
-        # Retrieve rows based on options
-        rows = self._get_rows(options)
-        # Convert rows to list of dicts
-        lines = [self.format_row(row) for row in rows]
-        # Serialize to JSON with indentation and sorted keys
-        return json.dumps(lines, indent=2, sort_keys=True)
-
-
-class JsonLicenseFinderTable(JsonPrettyTable):
-
-    def format_row(self, row: RowType) -> dict[str, str | list[str]]:
-        resrow: dict[str, str | list[str]] = {}
-        for field, value in zip(self._field_names, row):
-            if field == "Name":
-                resrow["name"] = value
-            if field == "Version":
-                resrow["version"] = value
-            if field == "License":
-                resrow["licenses"] = [value]
-        return resrow
-
-    def get_string(self, **kwargs: str | list[str]) -> str:
-        # import included here in order to limit dependencies
-        # if not interested in JSON output,
-        # then the dependency is not required
-        import json
-
-        options = self._get_options(kwargs)
-        rows = self._get_rows(options)
-        lines = [self.format_row(row) for row in rows]
-        return json.dumps(lines, sort_keys=True)
