--- piplicenses/output/plain.py
@@ -1,48 +1,0 @@
-# pip-licenses-cli
-#
-# MIT License
-#
-# Copyright (c) 2018 raimon
-# Copyright (c) 2025 stefan6419846
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in
-# all copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
-from __future__ import annotations
-
-from prettytable import PrettyTable
-
-
-class PlainVerticalTable(PrettyTable):
-    """PrettyTable for outputting to a simple non-column based style.
-
-    When used with --with-license-file, this style is similar to the default
-    style generated from Angular CLI's --extractLicenses flag.
-    """
-
-    def get_string(self, **kwargs: str | list[str]) -> str:
-        options = self._get_options(kwargs)
-        rows = self._get_rows(options)
-
-        output = ""
-        for row in rows:
-            for v in row:
-                output += "{}\n".format(v)
-            output += "\n"
-
-        return output
